requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
Faker==30.3.0
google==3.0.0
google-api-core==2.14.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
idna==3.6
jwcrypto==1.5.0
oauthlib==3.2.2
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wrapt==1.16.0